chore: move nav interface, remove unused constructor
